7주차 과제 풀이 - advanced
