+++ packages/assignment-7/src/shared/categories.ts
@@ -1,0 +1,1 @@
+export const categories = ["업무", "개인", "가족", "기타"];
