+++ packages/assignment-7/src/types/event.ts
@@ -1,0 +1,20 @@
+type RepeatType = "none" | "daily" | "weekly" | "monthly" | "yearly";
+
+interface RepeatInfo {
+  type: RepeatType;
+  interval: number;
+  endDate?: string;
+}
+
+export interface EventType {
+  id: number;
+  title: string;
+  date: string;
+  startTime: string;
+  endTime: string;
+  description: string;
+  location: string;
+  category: string;
+  repeat: RepeatInfo;
+  notificationTime: number; // 분 단위로 저장
+}
